Change Access Modifier & `BaseSettingsPage` Handles `mouseClicked` & Made Button Actions Betterer & Cleanup
