+++ src/main/java/io/redstudioragnarok/fbp/gui/menu/BaseSettingsPage.java
@@ -16,41 +16,42 @@ import static io.redstudioragnarok.fbp.gui.FBPGuiButton.ButtonSize.*;
 
 public abstract class BaseSettingsPage extends GuiScreen {
 
-    private GuiScreen previousPage, nextPage;
+    protected int x;
 
     protected GuiButton toggle, issue;
 
-    @Override
-    public void initGui() {
-        int x = width / 2 - 100;
+    private GuiScreen previousPage, nextPage;
+
+    // Todo: Actually use this and set it to true in the actionPerformed method & check against it in onGuiClosed
+    private boolean writeConfig;
+
+
+    public void initPage(GuiScreen previousPage, GuiScreen nextPage) {
+        x = width / 2 - 100;
+
         int xTop = width - 27;
 
-        int y = height / 5 + 149;
-        int yNavigation = height / 5 + 101;
+        int y = height / 5 + 101;
 
         addButton(0, x, y, medium, I18n.format("menu.done"), false, false, true);
         addButton(-1, x + 102, y, medium, I18n.format("menu.defaults"), false, false, true);
 
         addButton(-2, x, y + 20 + 1, large, I18n.format("menu.reloadconfig"), false, false, true);
 
+        // Todo: Use the global X instead of xTop variable and use math to get it to the same pos
         toggle = new GuiButtonEnable(-3, xTop - 29, 2, this.fontRenderer);
         issue = new GuiButtonBugReport(-4, xTop, 2, new Dimension(width, height), this.fontRenderer);
 
         this.buttonList.addAll(Arrays.asList(toggle, issue));
-    }
 
-    public void initNavigation(GuiScreen previousPage, GuiScreen nextPage) {
         this.previousPage = previousPage;
         this.nextPage = nextPage;
 
-        int x = width / 2 - 145;
-        int y = height / 5 + 101;
-
-        if (previousPage != null)
-            addButton(-5, x, y, small, "\u00A76<<", false, false, true);
+        if (previousPage!= null)
+            addButton(-5, x - 45, y, small, "\u00A76<<", false, false, true);
 
         if (nextPage!= null)
-            addButton(-6, x + 270, y, small, "\u00A76>>", false, false, true);
+            addButton(-6, x + 225, y, small, "\u00A76>>", false, false, true);
     }
 
     @Override
@@ -82,6 +83,11 @@ public abstract class BaseSettingsPage extends GuiScreen {
                 mc.displayGuiScreen(nextPage);
                 break;
         }
+
+        onActionPerformed(button);
+    }
+
+    protected void onActionPerformed(GuiButton button) {
     }
 
     @Override
@@ -102,19 +108,38 @@ public abstract class BaseSettingsPage extends GuiScreen {
         super.drawScreen(mouseX, mouseY, partialTicks);
     }
 
-    public FBPGuiButton addButton(int id, int x, int y, FBPGuiButton.ButtonSize size, String text, Boolean toggle, Boolean toggleButton, Boolean enabled) {
+    @Override
+    protected void mouseClicked(int mouseX, int mouseY, int mouseButton) {
+        if (mouseButton == 0) {
+            for (GuiButton guibutton : this.buttonList) {
+                if (guibutton.mousePressed(this.mc, mouseX, mouseY)) {
+                    if (!guibutton.isMouseOver())
+                        return;
+
+                    actionPerformed(guibutton);
+                }
+            }
+        }
+    }
+
+    @Override
+    public void onGuiClosed() {
+        ConfigHandler.writeMainConfig();
+    }
+
+    protected FBPGuiButton addButton(int id, int x, int y, FBPGuiButton.ButtonSize size, String text, Boolean toggle, Boolean toggleButton, Boolean enabled) {
         FBPGuiButton button = new FBPGuiButton(id, x, y, size, text, toggle, toggleButton, enabled);
         this.buttonList.add(button);
         return button;
     }
 
-    public FBPGuiButton addButton(int id, int x, int y, String text, Boolean toggle, Boolean toggleButton, Boolean enabled) {
+    protected FBPGuiButton addButton(int id, int x, int y, String text, Boolean toggle, Boolean toggleButton, Boolean enabled) {
         FBPGuiButton button = new FBPGuiButton(id, x, y, text, toggle, toggleButton, enabled);
         this.buttonList.add(button);
         return button;
     }
 
-    public GuiSlider addSlider(int x, int y, float value) {
+    protected GuiSlider addSlider(int x, int y, float value) {
         GuiSlider slider = new GuiSlider(x, y, value);
         this.buttonList.add(slider);
         return slider;
